big_alloc() now takes a 64-bit argument.
We would like to allocate buffers larger than 4 GB on 64-bit machines.
I've purposefully used u64 instead of size_t, because configuration
variables for buffer sizes will be usually u64's and big_alloc() is
the proper place to report buffer size errors.

+++ lib/lib.h
@@ -265,7 +265,7 @@ byte *str_unesc(byte *dest, byte *src);
 
 /* bigalloc.c */
 
-void *big_alloc(unsigned int len);
-void big_free(void *start, unsigned int len);
+void *big_alloc(u64 len);
+void big_free(void *start, u64 len);
 
 #endif
